+++ functions/trigger_manual_approval/triggerManualApproval.py
@@ -21,7 +21,10 @@ import os
 from botocore.exceptions import ClientError
 
 '''
-Get number of findings from Macie classification job. 
+Tag all objects and move to the manual review S3 bucket if sensitive data 
+found. Move files to the scanned data S3 bucket if no sensitive data found. 
+Trigger a manual approval notification to SNS topic if sensitive data is 
+found.
 
 This proof of concept is used as part of a data pipeline workflow as part of
 the data ingestion pipeline. 
@@ -37,7 +40,7 @@ def lambda_handler(event, context):
     src_bucket_name = os.environ['sourceS3Bucket']
     target_scanned_bucket_name = os.environ['targetScannedS3Bucket']
 
-
+    prefix = event['Input']['id']
     s3_key_names = event['Input']['macieFindingsInfo']['Payload']
     
     sns_client = boto3.client('sns')
@@ -92,20 +95,32 @@ def lambda_handler(event, context):
                 s3_keys_remaining = []
                 
             for s3_key_info in s3_keys_remaining:
-                print(f"Moving: {s3_key_info['Key']}")
-                response = s3_client.copy_object(
-                    Bucket = target_scanned_bucket_name,
-                    CopySource = {
-                        'Bucket': src_bucket_name,
-                        'Key': s3_key_info['Key']
-                    },
-                    Key = s3_key_info['Key']
-                )
-                print('Deleting object')
-                response = s3_client.delete_object(
+                print(f"Checking for tag on: {s3_key_info['Key']}")
+                object_tags = s3_client.get_object_tagging(
                     Bucket=src_bucket_name,
-                    Key = s3_key_info['Key']
+                    Key=s3_key_info['Key']
                 )
+                for tag_set in object_tags['TagSet']:
+                    if tag_set['Key'] == 'WorkflowId':
+                        check_key = tag_set['Value']
+
+                if check_key == prefix:
+                    print(f"Moving: {s3_key_info['Key']}")
+                    response = s3_client.copy_object(
+                        Bucket = target_scanned_bucket_name,
+                        CopySource = {
+                            'Bucket': src_bucket_name,
+                            'Key': s3_key_info['Key']
+                        },
+                        Key = s3_key_info['Key']
+                    )
+                    print('Deleting object')
+                    response = s3_client.delete_object(
+                        Bucket=src_bucket_name,
+                        Key = s3_key_info['Key']
+                    )
+                else:
+                    print(f"Object tag not matching for {s3_key_info['Key']}")
     except Exception as e:
         print(f'Could not complete S3 object move')
         print(e)
